requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-genai==1.63.0
